+++ source/minitest/debug.c
@@ -1,0 +1,11 @@
+void debug_print(t_label *lb)
+{
+	t_label *tmp;
+
+	tmp = lb;
+	while (tmp)
+	{
+		printf("L:\t%s\n", tmp->name);
+		tmp = tmp->next;
+	}
+}
